Revert to old MockQuoteSorter
