+++ KarhooUISDKTests/TestCases/Screens/BookingScreen/QuoteList/KarhooQuoteListPresenterSpec.swift
@@ -268,14 +268,3 @@ final class KarhooQuoteListPresenterSpec: XCTestCase {
         XCTAssertTrue(mockQuoteListView.showQuotesAnimated!)
     }
 }
-
-class MockQuoteSorter: QuoteSorter {
-    
-    private(set) var sortQuotesSet: [Quote]?
-    private(set) var orderSet: QuoteSortOrder?
-
-    var quotesToReturn: [Quote] = []
-    func sortQuotes(_ quotes: [Quote], by order: QuoteSortOrder) -> [Quote] {
-        return quotesToReturn
-    }
-}
